+++ lib/editor/page.ex
@@ -28,14 +28,20 @@ defmodule Editor.Page do
   @spec insert_block(t, block_id :: id, cell_id :: id, integer) :: t
   def insert_block(%__MODULE__{blocks: blocks} = page, block_id, cell_id, index) do
     current_block_index = Enum.find_index(blocks, &(&1.id === block_id))
-    current_block = Enum.at(blocks, current_block_index)
+    %Editor.Block{} = current_block = Enum.at(blocks, current_block_index)
 
     case Editor.Block.split(current_block, cell_id, index) do
-      [%{} = block] ->
-        blocks = List.replace_at(blocks, current_block_index, block)
-        %{page | blocks: blocks, active_cell_id: nil, cursor_index: nil}
+      [%Editor.Block{} = new_block] ->
+        old_cell_ids = Enum.map(current_block.cells, & &1.id)
+        new_cell_ids = Enum.map(new_block.cells, & &1.id)
+        new_cell_id = Enum.find(new_cell_ids, &(&1 not in old_cell_ids))
+        %Editor.Cell{} = new_cell = Enum.find(new_block.cells, &(&1.id === new_cell_id))
+        active_cell_id = new_cell.id
+        cursor_index = String.length(new_cell.content)
+        blocks = List.replace_at(blocks, current_block_index, new_block)
+        %{page | blocks: blocks, active_cell_id: active_cell_id, cursor_index: cursor_index}
 
-      [_, new_block] = new_blocks ->
+      [_, %Editor.Block{} = new_block] = new_blocks ->
         blocks = blocks |> List.replace_at(current_block_index, new_blocks) |> List.flatten()
         active_cell = Enum.at(new_block.cells, 0)
         %{page | blocks: blocks, active_cell_id: active_cell.id, cursor_index: 0}
@@ -45,15 +51,21 @@ defmodule Editor.Page do
   @spec update_block(t, block_id :: id, cell_id :: id, String.t()) :: t
   def update_block(%__MODULE__{blocks: blocks} = page, block_id, cell_id, value) do
     block_index = Enum.find_index(blocks, &(&1.id === block_id))
+    %Editor.Block{} = old_block = Enum.at(blocks, block_index)
+    %Editor.Block{} = new_block = Editor.Block.update(old_block, cell_id, value)
 
-    block =
-      blocks
-      |> Enum.at(block_index)
-      |> Editor.Block.update(cell_id, value)
+    blocks = List.replace_at(blocks, block_index, new_block)
 
-    blocks = List.replace_at(blocks, block_index, block)
+    %Editor.Cell{} = cell = Enum.find(new_block.cells, &(&1.id === cell_id))
 
-    %{page | blocks: blocks, active_cell_id: cell_id, cursor_index: nil}
+    cursor_index =
+      if old_block.type != new_block.type do
+        String.length(cell.content)
+      else
+        nil
+      end
+
+    %{page | blocks: blocks, active_cell_id: cell_id, cursor_index: cursor_index}
   end
 
   @spec backspace(t, block_id :: id, cell_id :: id) :: t
@@ -102,5 +114,22 @@ defmodule Editor.Page do
             cursor_index: String.length(old_cell.content)
         }
     end
+  end
+
+  @spec paste_blocks(t, list(Editor.Block.t()), String.t(), String.t(), integer) :: t
+  def paste_blocks(%__MODULE__{} = page, blocks, block_id, cell_id, index)
+      when is_list(blocks) and is_binary(block_id) and is_binary(cell_id) and is_integer(index) do
+    current_block_index = Enum.find_index(page.blocks, &(&1.id === block_id))
+    current_block = Enum.at(page.blocks, current_block_index)
+
+    clones = Enum.map(blocks, &Editor.Block.clone/1)
+
+    [part_before, part_after] = Editor.Block.hard_split(current_block, cell_id, index)
+
+    new_blocks = [part_before] ++ clones ++ [part_after]
+    all_blocks = page.blocks |> List.replace_at(current_block_index, new_blocks) |> List.flatten()
+    active_cell = Enum.at(part_after.cells, 0)
+
+    %{page | blocks: all_blocks, active_cell_id: active_cell.id, cursor_index: 0}
   end
 end