+++ lib/philtre_web/live/article_live/new.ex
@@ -8,25 +8,24 @@ defmodule PhiltreWeb.ArticleLive.New do
 
   @spec mount(map, %LiveView.Session{}, LiveView.Socket.t()) :: {:ok, LiveView.Socket.t()}
   def mount(%{}, _session, socket) do
-    {:ok, assign(socket, :page, Editor.Page.new())}
+    {:ok, assign(socket, :editor, Editor.new())}
   end
 
+  @errors_saving "There were some errors saving the article"
+
   def handle_event("save", %{}, socket) do
-    %Editor.Page{} = page = socket.assigns.page
+    %Editor.Page{} = page = socket.assigns.editor.page
 
     socket =
       case Articles.create_article(page) do
-        {:ok, _article} ->
-          push_redirect(socket, to: "/articles")
-
-        {:error, _changeset} ->
-          put_flash(socket, :error, "There were some errors saving the article")
+        {:ok, _article} -> push_redirect(socket, to: "/articles")
+        {:error, _changeset} -> put_flash(socket, :error, @errors_saving)
       end
 
     {:noreply, socket}
   end
 
-  def handle_info({:updated_page, page}, socket) do
-    {:noreply, assign(socket, :page, page)}
+  def handle_info({:update, %Editor{} = editor}, socket) do
+    {:noreply, assign(socket, :editor, editor)}
   end
 end